+++ inc/LNS/Parallel/GlobalManager.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "LNS/Parallel/DataStructure.h"
 #include "LNS/Parallel/NeighborGenerator.h"
 #include "LNS/Parallel/LocalOptimizer.h"
 #include "util/TimeLimiter.h"
@@ -14,11 +15,11 @@ public:
     std::shared_ptr<NeighborGenerator> neighbor_generator;
     std::vector<std::shared_ptr<LocalOptimizer>> local_optimizers;
 
-    int initial_sum_of_costs=MAX_COST;
-    int sum_of_costs=MAX_COST;
+    float initial_sum_of_costs=MAX_COST;
+    float sum_of_costs=MAX_COST;
     int num_of_failures=0;
     double average_group_size=0;
-    int sum_of_distances = 0;
+    float sum_of_distances = 0;
     int window_size_for_CT;
     int window_size_for_CAT;
     int window_size_for_PATH;
@@ -29,12 +30,12 @@ public:
     PathTable path_table;
     std::vector<Agent> agents;
     std::shared_ptr<HeuristicTable> HT;
-    std::shared_ptr<vector<int> > map_weights;
+    std::shared_ptr<vector<float> > map_weights;
     int screen;
     int num_threads;
 
     GlobalManager(
-        Instance & instance, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<int> > map_weights,
+        Instance & instance, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<float> > map_weights,
         int neighbor_size, destroy_heuristic destroy_strategy,
         bool ALNS, double decay_factor, double reaction_factor,
         string init_algo_name, string replan_algo_name, bool sipp,
