+++ inc/LNS/Parallel/LocalOptimizer.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "LNS/PathTable.h"
-#include "LNS/BasicLNS.h"
 #include "util/HeuristicTable.h"
 #include "util/TimeLimiter.h"
 #include "LNS/Parallel/TimeSpaceAStarPlanner.h"
@@ -17,7 +16,7 @@ public:
     PathTable path_table; // maintain a copy
     std::vector<Agent> & agents; // remove in the future, currently we can visit it for agent id but not do anything else.
     std::shared_ptr<HeuristicTable> HT;
-    std::shared_ptr<vector<int> > map_weights;
+    std::shared_ptr<vector<float> > map_weights;
     std::shared_ptr<TimeSpaceAStarPlanner> path_planner;
 
     string replan_algo_name;
@@ -28,7 +27,7 @@ public:
     int screen=0;
 
     LocalOptimizer(
-        Instance & instance, std::vector<Agent> & agents, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<int> > map_weights,
+        Instance & instance, std::vector<Agent> & agents, std::shared_ptr<HeuristicTable> HT, std::shared_ptr<vector<float> > map_weights,
         string replan_algo_name, bool sipp,
         int window_size_for_CT, int window_size_for_CAT, int window_size_for_PATH,
         int screen
