+++ inc/LaCAM2/post_processing.hpp
@@ -14,9 +14,9 @@ bool is_feasible_solution(const Instance& ins, const Solution& solution,
 int get_makespan(const Solution& solution);
 int get_path_cost(const Solution& solution, uint i);  // single-agent path cost
 int get_sum_of_costs(const Solution& solution);
-int get_sum_of_loss(const Solution& solution);
-int get_makespan_lower_bound(const Instance& ins, const std::shared_ptr<HeuristicTable>& HT);
-int get_sum_of_costs_lower_bound(const Instance& ins, const std::shared_ptr<HeuristicTable>& HT);
+float get_sum_of_loss(const Solution& solution);
+float get_makespan_lower_bound(const Instance& ins, const std::shared_ptr<HeuristicTable>& HT);
+float get_sum_of_costs_lower_bound(const Instance& ins, const std::shared_ptr<HeuristicTable>& HT);
 void print_stats(const int verbose, const Instance& ins, const std::shared_ptr<HeuristicTable>& HT,
                  const Solution& solution, const double comp_time_ms);
 void make_log(const Instance& ins, const std::shared_ptr<HeuristicTable>& HT, const Solution& solution,
